% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_max_tokens.R
\name{determine_max_tokens}
\alias{determine_max_tokens}
\title{Determine the maximum number of tokens based on the model}
\usage{
determine_max_tokens(model)
}
\description{
Determine the maximum number of tokens based on the model
}
