% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cost.R
\name{estimate_cost}
\alias{estimate_cost}
\title{Estimate the OpenAI API usage cost}
\usage{
estimate_cost(path = NULL, model = "gpt-3.5-turbo-16k")
}
\arguments{
\item{path}{The search path for the PDF}

\item{model}{The model to be used. This parameter is sent to the OpenAI API, so it should match with one of their \href{https://platform.openai.com/docs/models/gpt-3-5}{listed models}. With the \code{quickreader} package, the following models are supported: \code{gpt-3.5-turbo}, \verb{gpt-3.5-turbo-16k}, and \code{gpt-4}.}
}
\description{
Estimate the cost of having a specific model process all the text for a given path.
}
\details{
Calculates the expected cost for processing all the text found in \code{path}. The cost is calculated based on \href{https://openai.com/pricing}{OpenAI's pricing site}.
}
