% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Create a document from a \code{quickread} object}
\usage{
export(
  x,
  filename = "quickread",
  format = "pdf",
  toc = {
     class(x) == "list"
 },
  number_sections = {
     class(x) == "list"
 },
  document_information = TRUE,
  print_summary = TRUE
)
}
\arguments{
\item{x}{A list of the class \code{quickread}.}
}
\description{
Create a document from a \code{quickread} object
}
