% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Create a pipe-style table for a \code{quickread} class object}
\usage{
info(
  x = NULL,
  path = x$path,
  ntokens = x$ntokens,
  summary = x$summary,
  instruction = x$instruction,
  model = x$model,
  max_tokens = x$max_tokens,
  type = "document"
)
}
\description{
Create a pipe-style table for a \code{quickread} class object
}
