% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntokens.R
\name{ntokens}
\alias{ntokens}
\title{Count the number of tokens}
\usage{
ntokens(text)
}
\arguments{
\item{text}{A string whose token encoding one wants to count.}
}
\description{
This function counts the number of tokens in a character vector called \code{text} using the python package \code{tiktoken} provided by OpenAI.
}
