% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwords.R
\name{nwords}
\alias{nwords}
\title{Count the number of words}
\usage{
nwords(text)
}
\arguments{
\item{text}{An input text (character vector with one element).}
}
\description{
This function counts the number of words in a character vector called \code{text}. Note that the function splits the text by white spaces, thus numbers and connected special character also count as words.
}
