% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2text.R
\name{pdf2text}
\alias{pdf2text}
\title{Convert a PDF file to text}
\usage{
pdf2text(path, remove = c("references", "brackets"))
}
\arguments{
\item{path}{The search path to a locally stored PDF file or the URL to a PDF in the internet.}

\item{remove}{A character vector specifying what to remove from the text. Currently supports \code{"references"} for removing the list of references from an article, \code{"brackets"} for removing all content within brackets, and \code{"parentheses"} for removing all content within parentheses.}
}
\description{
This function allows reading a PDF file into R either from the local drive or from the web and extracting the text from that PDF.
}
