% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.quickread.R
\name{print.quickread}
\alias{print.quickread}
\title{Print out the response of a quickread object}
\usage{
\method{print}{quickread}(x)
}
\arguments{
\item{x}{A list of the class \code{quickread}.}
}
\description{
Print out the response of a quickread object
}
