% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressbar.R
\name{progressbar}
\alias{progressbar}
\title{Printing a progress bar}
\usage{
progressbar(i, n, message = "")
}
\arguments{
\item{i}{The current step.}

\item{n}{The total number of steps, must be greater than i.}

\item{message}{An optional message to be printed in front of the progress bar.}
}
\description{
This function prints out a progress bar in the console.
}
