% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickread.R
\name{quickread}
\alias{quickread}
\title{Send a prompt to ChatGPT}
\usage{
quickread(
  path = NULL,
  instruction = NULL,
  enable_shorten = TRUE,
  language = "English",
  remove = c("references", "brackets"),
  verbose = TRUE,
  model = "gpt-3.5-turbo-16k",
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  max_tokens = determine_max_tokens(model)
)
}
\arguments{
\item{path}{The search path for the PDF}

\item{instruction}{The instruction for ChatGPT}

\item{enable_shorten}{If the text extracted from the PDF exceeds \code{max_tokens}, should the text be split up in chunks, individually summarized and be put together again? Defaults to \code{TRUE}. Note that this allows heavy (charged) usage of ChatGPT, and in the case of \code{model = "gpt-4"}, it might lead to an error due to OpenAI's usage limit for its models.}

\item{language}{Language for both the hidden ChatGPT prompts as well as the document output. Currently supports \code{"English"} and \code{"German"}. Note that the GPT models have been trained on data from the internet, which includes text in hundreds of languages, but is very biased towards English. Thus, the model's responses and "reasoning skills" might be the best in English, even though it reads a PDF in another language.}

\item{remove}{A character vector specifying what to remove from the text. Currently supports \code{"references"} for removing the list of references from an article, \code{"brackets"} for removing all content within brackets, and \code{"parentheses"} for removing all content within parentheses.}

\item{verbose}{Should progress be printed in the console?}

\item{model}{The model to be used. This parameter is sent to the OpenAI API, so it should match with one of their \href{https://platform.openai.com/docs/models/gpt-3-5}{listed models}. With the \code{quickreader} package, the following models are supported: \code{gpt-3.5-turbo}, \verb{gpt-3.5-turbo-16k}, and \code{gpt-4}.}

\item{openai_api_key}{The API key for the OpenAI API. By default, the key will be searched for as \code{OPENAI_API_KEY} in the system environments. In order to set up an OpenAI API key, you need to sign up for the \href{https://openai.com/product}{OpenAI API}, and generate your personal secret API key on \href{https://platform.openai.com/}{this site} (Personal, and select View API keys in drop-down menu). You can then copy the key by clicking on the green text Copy.}

\item{max_tokens}{The maximal number of tokens the text sent to ChatGPT may consist of. If the number of tokens exceeds \code{max_token}, the text is split up and iteratively processed.}

\item{description}{TODO}

\item{details}{TODO}
}
\description{
Send a prompt to ChatGPT
}
