% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGPT.R
\name{runGPT}
\alias{runGPT}
\title{Send a prompt to ChatGPT}
\usage{
runGPT(
  prompt,
  model = "gpt-3.5-turbo-16k",
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{prompt}{The prompt to be sent to ChatGPT}

\item{model}{The model to be used. This parameter is sent to the OpenAI API, so it should match with one of their \href{https://platform.openai.com/docs/models/gpt-3-5}{listed models}.}

\item{openai_api_key}{The API key for the OpenAI API. By default, the key will be searched for as \code{OPENAI_API_KEY} in the system environments.}
}
\description{
Send a prompt to ChatGPT
}
