% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten.R
\name{shorten}
\alias{shorten}
\title{Shorten a too long text}
\usage{
shorten(
  text,
  path = NULL,
  language = "English",
  verbose = TRUE,
  max_tokens = 14000,
  model = "gpt-3.5-turbo-16k",
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{text}{An input text (character vector with one element).}

\item{max_tokens}{The maximal number of tokens one sub-part of the text ought to consist of.}
}
\description{
This function shortens texts that are too long for OpenAI's LLMs to handle at once by splitting it up in chunks and summarizing them individually.
}
