% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.quickread.R
\name{summary.quickread}
\alias{summary.quickread}
\title{Print out the summary of a text, if present}
\usage{
\method{summary}{quickread}(x)
}
\arguments{
\item{x}{A list of the class \code{quickread}.}
}
\description{
Print out the summary of a text, if present
}
