% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textsplit.R
\name{textsplit}
\alias{textsplit}
\title{Split the text at given sentences}
\usage{
textsplit(text, max_tokens = 12000)
}
\arguments{
\item{text}{An input text (character vector with one element).}

\item{max_tokens}{The maximal number of tokens one sub-part of the text ought to consist of.}
}
\description{
This function splits an input text into \code{n} parts, where the sum of tokens of every part is less than or equal to \code{max_tokens}. It returns a character vector with the individual text parts.
}
\details{
For better comprehension of the return value of \code{textsplit}, the text is only split by full sentences and not by individual tokens.
}
